making private components and extracted public components
