+++ app/src/main/java/com/example/greetingcard/view/LoginScreen.kt
@@ -6,11 +6,8 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -20,12 +17,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
+import com.example.greetingcard.view.component.InputFields
 import com.example.greetingcard.viewModel.LoginViewModel
 
 @Composable
@@ -87,46 +83,23 @@ fun LoginScreen(navController: NavHostController) {
 }
 
 @Composable
-fun Title() {
+private fun Title() {
     Text(
         fontWeight = FontWeight.Bold, fontSize = 24.sp, text = "Bem vindo(a) à Taqtile!"
     )
 }
 
-@Composable
-fun InputFields(
-    value: String,
-    title: String,
-    type: KeyboardType,
-    onValueChange: (String) -> Unit,
-    isPassword: Boolean = false
-) {
-    Column() {
-        Text(text = title)
-        OutlinedTextField(
-            value = value,
-            onValueChange = onValueChange,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 16.dp),
-            keyboardOptions = KeyboardOptions(
-                keyboardType = type
-            ),
-            shape = RoundedCornerShape(16.dp),
-            visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None
-        )
-    }
-}
+
 
 @Composable
-fun LoginButton(onClick: () -> Unit) {
+private fun LoginButton(onClick: () -> Unit) {
     Button(onClick = onClick, modifier = Modifier.fillMaxWidth()) {
         Text(fontSize = 20.sp, text = "Entrar")
     }
 }
 
 @Composable
-fun ShowErrors(errorMessages: List<String>) {
+private fun ShowErrors(errorMessages: List<String>) {
     if (errorMessages.isNotEmpty()) {
         for (error in errorMessages) {
             Text(
