+++ app/src/main/java/com/example/greetingcard/view/component/InputFields.kt
@@ -1,0 +1,40 @@
+package com.example.greetingcard.view.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun InputFields(
+    value: String,
+    title: String,
+    type: KeyboardType,
+    onValueChange: (String) -> Unit,
+    isPassword: Boolean = false
+) {
+    Column() {
+        Text(text = title)
+        OutlinedTextField(
+            value = value,
+            onValueChange = onValueChange,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 16.dp),
+            keyboardOptions = KeyboardOptions(
+                keyboardType = type
+            ),
+            shape = RoundedCornerShape(16.dp),
+            visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None
+        )
+    }
+}
